[cf_auth] : :pencil2: fix filename + ClassName
